+++ docker/config/test.env
@@ -32,6 +32,8 @@ OIDC_OP_USER_ENDPOINT=https://auth.example.com/authorize
 
 SYMBOL_FILE_PREFIX=v0
 
+UPLOAD_TEMPDIR=/tmp/test/uploads
+
 
 # Eliot settings
 # --------------
