+++ tecken/tests/test_decorators.py
@@ -2,10 +2,6 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this
 # file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-import os
-
-import pytest
-
 from django import http
 from django.test import RequestFactory
 
@@ -49,27 +45,3 @@ def test_set_cors_headers(rf):
     response = view_function(request)
     assert response["Access-Control-Allow-Origin"] == "example.com"
     assert response["Access-Control-Allow-Methods"] == "HEAD,GET"
-
-
-def test_make_tempdir():
-    mutable = []
-
-    @decorators.make_tempdir(prefix="PREFIX", suffix="SUFFIX")
-    def view_function(request, tmpdir, foo=None):
-        assert os.path.isdir(tmpdir)
-        basename = os.path.basename(tmpdir)
-        assert "PREFIX" in basename
-        assert "SUFFIX" in basename
-        mutable.append(tmpdir)
-
-        if foo == "ERROR":
-            raise NameError("anything")
-
-    view_function("somerequest", foo="bar")
-    assert not os.path.isdir(mutable[0])
-
-    # Let it fail this time
-    with pytest.raises(NameError):
-        view_function("somerequest", foo="ERROR")
-
-    assert not os.path.isdir(mutable[1])
